preliminary integration of TezosReporter into Reporter
- mv needed TezosReporter logic into Reporter
- revert re-org changes (EthReporter -> Reporter)
- revert changes to deployer tests
