+++ packages/deployer/test/deployer.js
@@ -2,7 +2,7 @@ const ganache = require("ganache-core");
 const Web3 = require("web3");
 const { createInterfaceAdapter } = require("@truffle/interface-adapter");
 const assert = require("assert");
-const { EthMigrationsV5: EthReporter } = require("@truffle/reporters");
+const Reporter = require("@truffle/reporters");
 const EventEmitter = require("events");
 
 const Deployer = require("../index");
@@ -76,7 +76,7 @@ describe("Deployer (sync)", function() {
       }
     };
     deployer = new Deployer(options);
-    reporter = new EthReporter();
+    reporter = new Reporter();
     reporter.setDeployer(deployer);
     reporter.setMigration(mockMigration);
     reporter.listen();
