+++ packages/migrate/index.js
@@ -3,10 +3,7 @@ const path = require("path");
 const glob = require("glob");
 const expect = require("@truffle/expect");
 const Config = require("@truffle/config");
-const {
-  EthMigrationsV5: EthReporter,
-  TezosMigrationsV5: TezosReporter
-} = require("@truffle/reporters");
+const Reporter = require("@truffle/reporters");
 const Migration = require("./migration.js");
 const Emittery = require("emittery");
 
@@ -21,9 +18,7 @@ const Migrate = {
   logger: null,
 
   launchReporter: function(config) {
-    if (config.networks[config.network].type === "tezos")
-      Migrate.reporter = new TezosReporter(config.describeJson || false);
-    else Migrate.reporter = new EthReporter(config.describeJson || false);
+    Migrate.reporter = new Reporter(config.describeJson || false);
     this.logger = config.logger;
   },
