+++ packages/reporters/index.js
@@ -1,7 +1,3 @@
-const EthMigrationsV5 = require("./reporters/EthMigrations-V5/reporter");
-const TezosMigrationsV5 = require("./reporters/TezosMigrationsV5/reporter");
+const Reporter = require("./reporters/migrations-V5/reporter");
 
-module.exports = {
-  EthMigrationsV5,
-  TezosMigrationsV5
-};
+module.exports = Reporter;
