--- packages/reporters/reporters/TezosMigrationsV5/indentedSpinner.js
@@ -1,15 +1,0 @@
-module.exports = {
-  interval: 80,
-  frames: [
-    "   ⠋",
-    "   ⠙",
-    "   ⠹",
-    "   ⠸",
-    "   ⠼",
-    "   ⠴",
-    "   ⠦",
-    "   ⠧",
-    "   ⠇",
-    "   ⠏"
-  ]
-};
