+++ packages/reporters/reporters/migrations-V5/indentedSpinner.js
